pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "computer-use-raw-python-executor"
version = "0.1.0"
description = "Unrestricted local Python executor for raw GUI-agent code."
readme = "README.md"
requires-python = ">=3.11"
dependencies = ["pillow"]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
computer-use-raw-python-executor = "computer_use_raw_python_executor.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
